refactor: move block, category, variable_assignment, node
