+++ package.nix
@@ -3,7 +3,7 @@ pkgs.rustPlatform.buildRustPackage {
   pname = "hyprlang-fmt";
   version = "0.1";
 
-  src = pkgs.lib.sourceByRegex ./. ["src" "src/cli" "src/components" "src/state" "src/pest" ".+\.rs" ".+\.pest" "^Cargo.lock$" "^Cargo.toml$"];
+  src = pkgs.lib.sourceByRegex ./. ["src" "src/cli" "src/components" "src/state" "src/grammar" ".+\.rs" ".+\.pest" "^Cargo.lock$" "^Cargo.toml$"];
 
   cargoHash = "sha256-+jB0Wpbj3vdPZ2mZU7muigbYuOzTONz8U0+cAz31VW0=";
