+++ src/components/mod.rs
@@ -1,2 +1,5 @@
 pub mod comment;
 pub mod command;
+pub mod variable_assignment;
+pub mod category;
+pub mod block;
